Added countdown to separate file
Moved the shuttle countdown function to its own file in the services folder instead of having it in the directions screen file to avoid overcrowding and promote function reuse.
