+++ screens/DirectionsScreen.tsx
@@ -11,7 +11,7 @@ import { SGW_CAMPUS } from "./MapExplorerScreen";
 import "react-native-get-random-values";
 import { useRoute } from "@react-navigation/native";
 import { retrieveRoutes } from "@/services/DirectionService.ts";
-import schedule from "@/assets/schedule.json";
+import { findNextShuttleCountdown } from "@/services/ShuttleService.ts"
 
 const googleMapsKey = GOOGLE_MAPS_API_KEY;
 const EDGE_PADDING = { top: 70, right: 70, bottom: 70, left: 70 };
@@ -119,28 +119,6 @@ export default function DirectionsScreen() {
     setTransportMode("DRIVING");  // buses drive, so driving mode
 
     await traceRoute(); // call trace route to trace shuttle bus service route
-  }
-
-  const findNextShuttleCountdown = (): string => {
-    const now = new Date();  // get current date time
-    const currentTime = now.getHours() * 60 + now.getMinutes(); // get time in minutes since day began
-
-    const shuttleSchedule = schedule.times;  // get shuttle times from schedule json
-
-    for (let i = 0; i < Object.keys(shuttleSchedule).length; i++) {
-        const timeArray = shuttleSchedule[i].split(":");  // split time string into hours and minutes
-        const hours = Number(timeArray[0]);
-        const minutes = Number(timeArray[1]);  // get hours and minutes in a number format
-        const time = hours * 60 + minutes;  // get the time in minutes for comparator
-
-        if (time > currentTime) {  // this means that the time coming up is the closest to our current time
-            // get minutes till shuttle by subtracting, turn to String and create message
-            const returnString = "Time Until Next Shuttle: " + String(time - currentTime) + " min";
-            return returnString;
-        }
-    }
-    const returnString = "No More Shuttles Today";
-    return returnString;  // fall back if no matches were found
   }
 
   useEffect (() => {
